multi-sig signer separated
Split the multi-sig test into two file, made on file specific only to signing and the private keys are abstracted in it. So far the logic for multi-signature is working out.
Still having problem on the broadcast side with error "NotEnoughFunds". Please ignore the comments in that part for right now.