Updated listening events
So I have sped up the transactions events, now it takes upto 20 seconds to process 450 transactions.

The result is now stored in files for later reviewing and caching.

So I made an extra function getTransactionEventsByTx which is similar to getTransactionEvents but it doesn't await just like that, so its much faster than it.

I think we should replace the existing one with the one I created because both serves the same purpose